[Rename] util과 response로 나눔
